+++ configuration.js
@@ -1,0 +1,18 @@
+const configuration = {
+  rounds: [
+    {
+      label: 'מחזור ראשון (2/7 - 13/7)',
+      full: false,
+    },
+    {
+      label: 'מחזור שני (16/7 - 27/7)',
+      full: false,
+    },
+    {
+      label: 'מחזור שלישי (30/7 - 10/8)',
+      full: false,
+    },
+  ],
+}
+
+export { configuration }